refactor: 移动 i18n.ts 到 common/ 目录
将国际化模块从 src/i18n.ts 移动到 src/common/i18n.ts。
旧路径保留 re-export 兼容层，更新了 12 处导入引用。
所有 352 个测试通过。
