+++ src/test/csmlogHoverProvider.test.ts
@@ -10,7 +10,7 @@
 
 import * as assert from 'assert';
 import * as path from 'path';
-import { __setLanguageOverrideForTests } from '../i18n';
+import { __setLanguageOverrideForTests } from '../common/i18n';
 
 // ---------------------------------------------------------------------------
 // Type stubs (matching vscode-mock.ts shapes)
